+++ curves/bw6_761/src/curves/g2.rs
@@ -19,7 +19,6 @@ impl CurveConfig for Config {
 
     /// COFACTOR =
     /// 26642435879335816683987677701488073867751118270052650655942102502312977592501693353047140953112195348280268661194869
-    #[rustfmt::skip]
     const COFACTOR: &'static [u64] = &[
         0x3de5800000000075,
         0x832ba4061000003b,
@@ -43,6 +42,12 @@ impl SWCurveConfig for Config {
 
     /// AFFINE_GENERATOR_COEFFS = (G2_GENERATOR_X, G2_GENERATOR_Y)
     const GENERATOR: G2Affine = G2Affine::new_unchecked(G2_GENERATOR_X, G2_GENERATOR_Y);
+
+    /// Correctness:
+    /// Substituting (0, 0) into the curve equation gives 0^2 = b.
+    /// Since b is not zero, the point (0, 0) is not on the curve.
+    /// Therefore, we can safely use (0, 0) as a flag for the zero point.
+    type ZeroFlag = ();
 
     #[inline(always)]
     fn mul_by_a(_elem: Self::BaseField) -> Self::BaseField {