+++ ec/src/scalar_mul/mod.rs
@@ -11,6 +11,7 @@ use ark_ff::{AdditiveGroup, BigInteger, PrimeField, Zero};
 use ark_std::{
     cfg_iter, cfg_iter_mut,
     ops::{Add, AddAssign, Mul, Neg, Sub, SubAssign},
+    vec,
     vec::*,
 };
 
@@ -21,7 +22,7 @@ use rayon::prelude::*;
 /// [`Explanation of usage`]
 ///
 /// [`Explanation of usage`]: https://github.com/scipr-lab/zexe/issues/79#issue-556220473
-fn ln_without_floats(a: usize) -> usize {
+const fn ln_without_floats(a: usize) -> usize {
     // log2(a) * ln(2)
     (ark_std::log2(a) * 69 / 100) as usize
 }
@@ -222,7 +223,7 @@ impl<T: ScalarMul> BatchMulPreprocessing<T> {
         }
     }
 
-    pub fn compute_window_size(num_scalars: usize) -> usize {
+    pub const fn compute_window_size(num_scalars: usize) -> usize {
         if num_scalars < 32 {
             3
         } else {