+++ ec/src/scalar_mul/variable_base/stream_pippenger.rs
@@ -99,13 +99,13 @@ impl<G: VariableBaseMSM> HashMapPippenger<G> {
             .or_insert(G::ScalarField::zero());
         *entry += *scalar.borrow();
         if self.buffer.len() == self.buf_size {
-            let bases = self.buffer.keys().cloned().collect::<Vec<_>>();
+            let bases = self.buffer.keys().copied().collect::<Vec<_>>();
             let scalars = self
                 .buffer
                 .values()
                 .map(|s| s.into_bigint())
                 .collect::<Vec<_>>();
-            self.result += G::msm_bigint(&bases, &scalars);
+            self.result += G::msm_bigint(&bases, scalars.as_slice());
             self.buffer.clear();
         }
     }
@@ -114,14 +114,14 @@ impl<G: VariableBaseMSM> HashMapPippenger<G> {
     #[inline(always)]
     pub fn finalize(mut self) -> G {
         if !self.buffer.is_empty() {
-            let bases = self.buffer.keys().cloned().collect::<Vec<_>>();
+            let bases = self.buffer.keys().copied().collect::<Vec<_>>();
             let scalars = self
                 .buffer
                 .values()
                 .map(|s| s.into_bigint())
                 .collect::<Vec<_>>();
 
-            self.result += G::msm_bigint(&bases, &scalars);
+            self.result += G::msm_bigint(&bases, scalars.as_slice());
         }
         self.result
     }