+++ test-templates/src/groups.rs
@@ -193,6 +193,16 @@ macro_rules! __test_group {
         }
 
         #[test]
+        fn test_var_base_msm_mixed_scalars() {
+            $crate::msm::test_var_base_msm_mixed_scalars::<$group>();
+        }
+
+        #[test]
+        fn test_var_base_msm_specialized() {
+            $crate::msm::test_var_base_msm_specialized::<$group>();
+        }
+
+        #[test]
         fn test_chunked_pippenger() {
             $crate::msm::test_chunked_pippenger::<$group>();
         }
@@ -278,6 +288,7 @@ macro_rules! __test_group {
 
         #[test]
         fn test_sw_properties() {
+            use ark_ec::models::short_weierstrass::Bucket;
             let mut rng = &mut ark_std::test_rng();
 
             let generator = <$group>::generator().into_affine();
@@ -305,6 +316,11 @@ macro_rules! __test_group {
                     }
                 }
             }
+            use core::any::TypeId;
+            if TypeId::of::<<Config as SWCurveConfig>::ZeroFlag>() == TypeId::of::<()>() {
+                let zero = BaseField::ZERO;
+                assert_ne!(zero.square() - zero.square() * zero - <Config as SWCurveConfig>::COEFF_A * zero - <Config as SWCurveConfig>::COEFF_B, zero);
+            }
 
             for _ in 0..ITERATIONS {
                 let f = BaseField::rand(rng);
@@ -327,6 +343,18 @@ macro_rules! __test_group {
                         assert_eq!(a, b);
                     }
 
+                }
+            }
+            {
+                for _ in 0..ITERATIONS {
+                    let a = Affine::rand(rng);
+                    let b = Affine::rand(rng);
+                    let mut a_bucket = Bucket::from(a);
+                    a_bucket += &b;
+                    assert_eq!(a + b, <$group>::from(a_bucket), "bucket + affine failed");
+                    let mut a_group = <$group>::from(a);
+                    a_group += &Bucket::from(b);
+                    assert_eq!(a + b, <$group>::from(a_group), "group + bucket failed");
                 }
             }
         }